+++ example/F0.Wpf.Example/F0.Wpf.Example.Windows/App.xaml.cs
@@ -1,0 +1,8 @@
+﻿using System.Windows;
+
+namespace F0.Wpf.Example.Windows
+{
+	internal partial class App : Application
+	{
+	}
+}