+++ example/F0.Wpf.Example/F0.Wpf.Example.Windows/MainWindow.xaml.cs
@@ -1,0 +1,12 @@
+﻿using System.Windows;
+
+namespace F0.Wpf.Example.Windows
+{
+	internal partial class MainWindow : Window
+	{
+		public MainWindow()
+		{
+			InitializeComponent();
+		}
+	}
+}